Add a test for the Jinja2 filter
