+++ .gitignore
@@ -7,4 +7,5 @@ dist
 # Test artefacts
 .tox
 .coverage
+.hypothesis
 .pytest_cache
