+++ tox.ini
@@ -4,6 +4,7 @@ envlist = py27, py34, py35, py36, lint
 [testenv]
 deps =
   coverage
+  hypothesis
   pytest
   pytest-cov
 commands =
